requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
openai==1.55.0
python-dotenv==1.0.0
httpx==0.25.2
playwright==1.40.0
Pillow==10.1.0
websockets==12.0
pyjwt==2.8.0

